Remove needlessly complex scoping rules (#3)
Originally, the design of our RDF connections was using Guice Scopes.
This effort was needlessly complicated (though I learned some things?).

This removes the scoping in favour of passing connections through.
Further, this differentiates between read and write connections more
directly. While in truth all connections are read-write, we should
separate the concerns should we want to optimize for reads in the
future.