Add plugin support for compression methods and rdf configurations (#11)
This creates a new class: `KAff4Plugin`.

Currently, this is a simple guice wrapper that has short-hands for
specific bindings.

### Considerations made

I considered using auto-registration via reflection, however that opens
up massive security concerns. This would imply that any plugin on the
classpath is automatically loaded. This is absolutely a non-starter and
implemented plugins can just as easily configure the injection point to
add new modules as needed. Should this be wanted, users can do this at
their own discretion.

### Iterations 

1. Add a "lifecycle" setup that allows for registering actions 
2. Add logging to list what plugins are loaded