[Logical] Add Query pagination to RDF (#24)
In order to list files, we need to do large queries against the RDF
store. These can be very large queries and we need to be mindful to only
query what we need.

This also strips out some unused methods from the APIs. 

Follow-up: Replace the usage of `Model` with some form of queries where
a connection is opened, utilized then closed. The current approach leads
to awkward usages where we open too many connections needlessly.