Implement Executors for contexts (#31)
This adds support for multithreaded verification via adding scoped
implementation of `ExecutorService`. This is done by providing a small
wrapper that copies the _current_ thread's action into each task and
creates a new scope to avoid polluting the context. This enables
multiple threads to share contexts as needed and avoids thread-local
issues.

### TODO
* [x] Implement threaded verification
* [x] Tests