Add tool dialect system (unused)
This adds a `ToolDialect` system for computing and registering behaviour that changes how the library parses different images.

This is necessitated due to defects, changes to standards, and general quirks with implementations that write AFF4 images. By leveraging these dialects,
we can isolate these behaviours in a consistent way and allow the library to adjust how it parses and reads these images.

This adds an `Aff4LogicalStandardToolDialect` which behaves as the `DefaultToolDialect` when none can be detected formally.

Extracted from #37