[Misk ActionScopes] Replace `guice-action-scoped` with `misk-action-scopes`
This performs the migration to swap out the custom built `guice-action-scoped` with `misk-action-scopes` library. Most of this change is switching the form `@ActionScoped fooProvider: Provider<Foo>` -> `fooProvider: ActionScoped<Foo>`. In #63, we updated the code paths to use `Provider<>` implementations which reduced the size of this PR substantially.

Extracted from #61, part of #55.