[Sources] Create kaff4-specific `Source` to use instead of `okio.Source` (#163)
This adds a new `Source` that replaces `okio.Source` in nearly all
cases. This is because we need one specific feature:

* Skipping for all sources: Each source implementation needs to
implement an efficient skip/forward-seek to make common usages efficient

This PR does not implement skipping for sources, instead it only does
the "replacement" with our own library version.

Depends on #162.